+++ package.json
@@ -5,28 +5,28 @@
     "node": ">=22.13.0"
   },
   "scripts": {
-    "server:dev": "dotenvx run -fk .env.keys -f server/.env -- pnpm --filter @aiss-nepch/server dev",
-    "server:deploy": "pnpm --filter @aiss-nepch/server run deploy",
-    "server:deploy:production": "pnpm --filter @aiss-nepch/server run deploy:production",
-    "web:dev": "dotenvx run -fk .env.keys -f web/.env -- pnpm --filter @aiss-nepch/web dev",
-    "web:build": "pnpm --filter @aiss-nepch/web build",
-    "web:deploy": "dotenvx run -fk .env.keys -f web/.env -- pnpm --filter @aiss-nepch/web run deploy",
-    "web:deploy:production": "dotenvx run -fk .env.keys -f web/.env.production -- pnpm --filter @aiss-nepch/web run deploy:production",
-    "mastra:dev": "dotenvx run -fk .env.keys -f server/.env -- pnpm --filter @aiss-nepch/server mastra:dev",
-    "mastra:build": "pnpm --filter @aiss-nepch/server mastra:build",
-    "mastra:start": "pnpm --filter @aiss-nepch/server mastra:start",
-    "lint": "biome check . && pnpm --filter @aiss-nepch/web exec tsc --noEmit && pnpm --filter @aiss-nepch/server exec tsc --noEmit",
+    "server:dev": "dotenvx run -fk .env.keys -f server/.env -- pnpm --filter @nepp-chan/server dev",
+    "server:deploy": "pnpm --filter @nepp-chan/server run deploy",
+    "server:deploy:production": "pnpm --filter @nepp-chan/server run deploy:production",
+    "web:dev": "dotenvx run -fk .env.keys -f web/.env -- pnpm --filter @nepp-chan/web dev",
+    "web:build": "pnpm --filter @nepp-chan/web build",
+    "web:deploy": "dotenvx run -fk .env.keys -f web/.env -- pnpm --filter @nepp-chan/web run deploy",
+    "web:deploy:production": "dotenvx run -fk .env.keys -f web/.env.production -- pnpm --filter @nepp-chan/web run deploy:production",
+    "mastra:dev": "dotenvx run -fk .env.keys -f server/.env -- pnpm --filter @nepp-chan/server mastra:dev",
+    "mastra:build": "pnpm --filter @nepp-chan/server mastra:build",
+    "mastra:start": "pnpm --filter @nepp-chan/server mastra:start",
+    "lint": "biome check . && pnpm --filter @nepp-chan/web exec tsc --noEmit && pnpm --filter @nepp-chan/server exec tsc --noEmit",
     "lint:fix": "biome check --write .",
     "format": "biome format --write .",
-    "db:generate": "pnpm --filter @aiss-nepch/server db:generate",
-    "db:migrate": "pnpm --filter @aiss-nepch/server db:migrate",
-    "db:migrate:production": "pnpm --filter @aiss-nepch/server db:migrate:production",
-    "db:migrate:local": "pnpm --filter @aiss-nepch/server db:migrate:local",
-    "db:studio": "pnpm --filter @aiss-nepch/server db:studio",
-    "db:push": "pnpm --filter @aiss-nepch/server db:push",
-    "db:check": "pnpm --filter @aiss-nepch/server db:check",
+    "db:generate": "pnpm --filter @nepp-chan/server db:generate",
+    "db:migrate": "pnpm --filter @nepp-chan/server db:migrate",
+    "db:migrate:production": "pnpm --filter @nepp-chan/server db:migrate:production",
+    "db:migrate:local": "pnpm --filter @nepp-chan/server db:migrate:local",
+    "db:studio": "pnpm --filter @nepp-chan/server db:studio",
+    "db:push": "pnpm --filter @nepp-chan/server db:push",
+    "db:check": "pnpm --filter @nepp-chan/server db:check",
     "knowledge:upload": "dotenvx run -fk .env.keys -f .env -- tsx scripts/upload-knowledge.ts",
-    "test": "pnpm --filter @aiss-nepch/server test"
+    "test": "pnpm --filter @nepp-chan/server test"
   },
   "devDependencies": {
     "@biomejs/biome": "^2.3.8",
