+++ server/src/mastra/index.ts
@@ -12,7 +12,7 @@ import { emergencyAgent } from "~/mastra/agents/emergency-agent";
 import { emergencyReporterAgent } from "~/mastra/agents/emergency-reporter-agent";
 import { feedbackAgent } from "~/mastra/agents/feedback-agent";
 import { knowledgeAgent } from "~/mastra/agents/knowledge-agent";
-import { nepChanAgent } from "~/mastra/agents/nepch-agent";
+import { nepChanAgent } from "~/mastra/agents/nepp-chan-agent";
 import { personaAgent } from "~/mastra/agents/persona-agent";
 import { personaAnalystAgent } from "~/mastra/agents/persona-analyst-agent";
 import { weatherAgent } from "~/mastra/agents/weather-agent";
@@ -75,7 +75,7 @@ export const mastra = new Mastra({
   observability: new Observability({
     configs: {
       default: {
-        serviceName: "nepch-agent",
+        serviceName: "nepp-chan-agent",
         sampling: { type: SamplingStrategyType.ALWAYS },
         exporters: [new DefaultExporter()],
       },
