+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@aiss-nepch/web",
+  "name": "@nepp-chan/web",
   "private": true,
   "version": "0.0.0",
   "type": "module",
