+++ web/src/lib/resource.ts
@@ -1,4 +1,4 @@
-const RESOURCE_ID_KEY = "nepch_resource_id";
+const RESOURCE_ID_KEY = "nepp_chan_resource_id";
 
 export const getResourceId = (): string => {
   let id = localStorage.getItem(RESOURCE_ID_KEY);
